Address PR comment: remove unnecessary axiom
